% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{brain_string_players_df}
\alias{brain_string_players_df}
\title{Brain Activity in String Players}
\format{
A data frame with 15 observations and 2 variables:
\describe{
  \item{Years}{Integer: Years of musical practice}
  \item{Activity}{Numeric: Brain activity measurement (likely fMRI or similar neuroimaging units)}
}
}
\source{
Data taken from the Sleuth3 package version 1.0-6. Original reference:
Ramsey, F.L. and Schafer, D.W. (2013) \emph{The Statistical Sleuth: A Course in Methods
of Data Analysis} (3rd ed), Cengage Learning.
}
\usage{
data(brain_string_players_df)
}
\description{
This dataset, brain_string_players_df, is a data frame containing neurophysiological
measurements from a study of 15 violin and other string instrument players.
The data examines the relationship between years of musical practice and
measured brain activity levels in relevant cortical regions.
}
\details{
The dataset name has been kept as 'brain_string_players_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the NeuroDataSets package. The suffix 'df' indicates
that the dataset is a data frame. The original content has not been modified.
}
