% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cocaine_dopamine_df}
\alias{cocaine_dopamine_df}
\title{Effects of Cocaine on Dopamine Receptors}
\format{
A data frame with 34 observations and 2 variables:
\describe{
  \item{percent.blocked}{Integer vector representing percent of dopamine receptors blocked}
  \item{high}{Integer vector representing perceived level of high from PET scans}
}
}
\source{
Data taken from the abd package version 0.2-8
}
\usage{
data(cocaine_dopamine_df)
}
\description{
This dataset, cocaine_dopamine_df, is a data frame containing measurements of
dopamine receptor blockade and perceived high levels from 34 human subjects
as determined by PET scans.
}
\details{
The dataset name has been kept as 'cocaine_dopamine_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
NeuroDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
