% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{neanderthal_brains_df}
\alias{neanderthal_brains_df}
\title{Cranial Capacity in Neanderthals and Modern Humans}
\format{
A data frame with 39 observations and 3 variables:
\describe{
  \item{ln.mass}{Numeric vector representing natural logarithm of body mass}
  \item{ln.brain}{Numeric vector representing natural logarithm of brain size}
  \item{species}{Factor indicating species with 2 levels (Neanderthals and early modern humans)}
}
}
\source{
Data taken from the abd package version 0.2-8
}
\usage{
data(neanderthal_brains_df)
}
\description{
This dataset, neanderthal_brains_df, is a data frame containing measurements of
brain size (lnbrain) and body mass (lnmass) from 39 specimens of Neanderthals
and early modern humans, identified by species.
}
\details{
The dataset name has been kept as 'neanderthal_brains_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
NeuroDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
