% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurodatasets-package.R
\docType{package}
\name{NeuroDataSets}
\alias{NeuroDataSets}
\alias{NeuroDataSets-package}
\title{NeuroDataSets: A Comprehensive Collection of Neuroscience and Brain-Related Datasets}
\description{
{
This package provides a diverse collection of datasets focused on the brain, nervous system, and related disorders. The package includes clinical, experimental, neuroimaging, behavioral, and cognitive data on conditions including Parkinson's, Alzheimer's, epilepsy, schizophrenia, autism, ADHD, Tourette's, TBI, brain tumors, migraines, sleep disorders, and mental health.
}
}
\details{
NeuroDataSets: A Comprehensive Collection of Neuroscience and Brain-Related Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Neuroscience and Brain-Related Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/neurodatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
