% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{dopamine_schizophrenia_tbl_df}
\alias{dopamine_schizophrenia_tbl_df}
\title{Dopamine \eqn{\beta}-Hydroxylase Activity in Schizophrenia}
\format{
A tibble with 25 observations and 2 variables:
\describe{
  \item{dbh}{Integer: Dopamine \eqn{\beta}-hydroxylase activity level (nmol/(mL\eqn{\cdot}hr))}
  \item{group}{Character: Treatment/patient group classification}
}
}
\source{
Data taken from the \pkg{BSDA} package version 1.2.2
}
\usage{
data(dopamine_schizophrenia_tbl_df)
}
\description{
This dataset, `dopamine_schizophrenia_tbl_df`, is a tibble containing measurements
of dopamine \eqn{\beta}-hydroxylase (DBH) activity in 25 schizophrenic patients treated
with antipsychotic medication. The data compares DBH levels between patient groups.
}
\details{
The dataset name has been kept as \code{dopamine_schizophrenia_tbl_df} to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the \pkg{NeuroDataSets} package. The suffix \code{tbl_df} indicates
that the dataset is a tibble. The original content has not been modified.
}
